feat: add example for develop connector
