--- examples/source/config.yaml
@@ -1,2 +1,0 @@
-target: "http://localhost:8080"
-source: "exampleSource"
