--- examples/source/secret.yaml
@@ -1,3 +1,0 @@
-host: "http://localhost:8080"
-username: "root"
-password: "123456"
